backend/api/menu/serializers.py
from rest_framework import serializers
from core.models import MenuItem

class MenuItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = MenuItem
        fields = ['id', 'name', 'description', 'price', 'available', 'category']
