backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework
djangorestframework-simplejwt
dotenv
drf_yasg